move resample_geotiff into image module
